N05 replaced uint with uint256

+++ contracts/Firewall.sol
@@ -163,19 +163,19 @@ contract Firewall is IFirewall, Ownable2Step {
      * @param data The calldata of the call (some firewall modifiers may pass custom data based on the use case)
      * @param value The value of the call.
      */
-    function preExecution(address sender, bytes calldata data, uint value) external override {
+    function preExecution(address sender, bytes calldata data, uint256 value) external override {
         bytes4 selector = bytes4(data);
         address[] memory policies = subscribedPolicies[msg.sender][selector];
         address[] memory globalPolicies = subscribedGlobalPolicies[msg.sender];
         if (dryrunEnabled[msg.sender]) {
-            for (uint i = 0; i < policies.length; i++) {
+            for (uint256 i = 0; i < policies.length; i++) {
                 try IFirewallPolicy(policies[i]).preExecution(msg.sender, sender, data, value) {
                     emit DryrunPolicyPreSuccess(msg.sender, selector, policies[i]);
                 } catch(bytes memory err) {
                     emit DryrunPolicyPreError(msg.sender, selector, policies[i], err);
                 }
             }
-            for (uint i = 0; i < globalPolicies.length; i++) {
+            for (uint256 i = 0; i < globalPolicies.length; i++) {
                 try IFirewallPolicy(globalPolicies[i]).preExecution(msg.sender, sender, data, value) {
                     emit DryrunPolicyPreSuccess(msg.sender, selector, globalPolicies[i]);
                 } catch(bytes memory err) {
@@ -183,10 +183,10 @@ contract Firewall is IFirewall, Ownable2Step {
                 }
             }
         } else {
-            for (uint i = 0; i < policies.length; i++) {
+            for (uint256 i = 0; i < policies.length; i++) {
                 IFirewallPolicy(policies[i]).preExecution(msg.sender, sender, data, value);
             }
-            for (uint i = 0; i < globalPolicies.length; i++) {
+            for (uint256 i = 0; i < globalPolicies.length; i++) {
                 IFirewallPolicy(globalPolicies[i]).preExecution(msg.sender, sender, data, value);
             }
         }
@@ -198,19 +198,19 @@ contract Firewall is IFirewall, Ownable2Step {
      * @param data The calldata of the call (some firewall modifiers may pass custom data based on the use case)
      * @param value The value of the call.
      */
-    function postExecution(address sender, bytes calldata data, uint value) external override {
+    function postExecution(address sender, bytes calldata data, uint256 value) external override {
         bytes4 selector = bytes4(data);
         address[] memory policies = subscribedPolicies[msg.sender][selector];
         address[] memory globalPolicies = subscribedGlobalPolicies[msg.sender];
         if (dryrunEnabled[msg.sender]) {
-            for (uint i = 0; i < policies.length; i++) {
+            for (uint256 i = 0; i < policies.length; i++) {
                 try IFirewallPolicy(policies[i]).postExecution(msg.sender, sender, data, value) {
                     emit DryrunPolicyPostSuccess(msg.sender, selector, policies[i]);
                 } catch(bytes memory err) {
                     emit DryrunPolicyPostError(msg.sender, selector, policies[i], err);
                 }
             }
-            for (uint i = 0; i < globalPolicies.length; i++) {
+            for (uint256 i = 0; i < globalPolicies.length; i++) {
                 try IFirewallPolicy(globalPolicies[i]).postExecution(msg.sender, sender, data, value) {
                     emit DryrunPolicyPostSuccess(msg.sender, selector, globalPolicies[i]);
                 } catch(bytes memory err) {
@@ -218,10 +218,10 @@ contract Firewall is IFirewall, Ownable2Step {
                 }
             }
         } else {
-            for (uint i = 0; i < policies.length; i++) {
+            for (uint256 i = 0; i < policies.length; i++) {
                 IFirewallPolicy(policies[i]).postExecution(msg.sender, sender, data, value);
             }
-            for (uint i = 0; i < globalPolicies.length; i++) {
+            for (uint256 i = 0; i < globalPolicies.length; i++) {
                 IFirewallPolicy(globalPolicies[i]).postExecution(msg.sender, sender, data, value);
             }
         }
@@ -238,7 +238,7 @@ contract Firewall is IFirewall, Ownable2Step {
     function preExecutionPrivateInvariants(
         address sender,
         bytes calldata data,
-        uint value
+        uint256 value
     ) external override returns (bytes32[] memory storageSlots) {
         bytes4 selector = bytes4(data);
         address privateInvariantsPolicy = subscribedPrivateInvariantsPolicy[msg.sender][selector];
@@ -269,7 +269,7 @@ contract Firewall is IFirewall, Ownable2Step {
     function postExecutionPrivateInvariants(
         address sender,
         bytes memory data,
-        uint value,
+        uint256 value,
         bytes32[] calldata preValues,
         bytes32[] calldata postValues
     ) external override {
@@ -352,7 +352,7 @@ contract Firewall is IFirewall, Ownable2Step {
      * accordingly.
      */
     function addGlobalPolicyForConsumers(address[] calldata consumers, address policy) external {
-        for (uint i = 0; i < consumers.length; i++) {
+        for (uint256 i = 0; i < consumers.length; i++) {
             require(msg.sender == IFirewallConsumer(consumers[i]).firewallAdmin(), "Firewall: not consumer admin");
             _addGlobalPolicy(consumers[i], policy);
         }
@@ -366,7 +366,7 @@ contract Firewall is IFirewall, Ownable2Step {
      * @param policy The address of the policy contract.
      */
     function removeGlobalPolicyForConsumers(address[] calldata consumers, address policy) external {
-        for (uint i = 0; i < consumers.length; i++) {
+        for (uint256 i = 0; i < consumers.length; i++) {
             require(msg.sender == IFirewallConsumer(consumers[i]).firewallAdmin(), "Firewall: not consumer admin");
             _removeGlobalPolicy(consumers[i], policy);
         }
@@ -388,7 +388,7 @@ contract Firewall is IFirewall, Ownable2Step {
      * accordingly.
      */
     function addPolicies(address consumer, bytes4[] calldata methodSigs, address[] calldata policies) external onlyConsumerAdmin(consumer) {
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             _addPolicy(consumer, methodSigs[i], policies[i]);
         }
     }
@@ -419,7 +419,7 @@ contract Firewall is IFirewall, Ownable2Step {
      * @param policies The addresses of the policy contracts.
      */
     function removePolicies(address consumer, bytes4[] calldata methodSigs, address[] calldata policies) external onlyConsumerAdmin(consumer) {
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             _removePolicy(consumer, methodSigs[i], policies[i]);
         }
     }
@@ -441,7 +441,7 @@ contract Firewall is IFirewall, Ownable2Step {
      * @param policies The addresses of the policy contracts.
      */
     function setPrivateInvariantsPolicy(address consumer, bytes4[] calldata methodSigs, address[] calldata policies) external onlyConsumerAdmin(consumer) {
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             require(approvedPolicies[policies[i]], "Firewall: policy not approved");
             subscribedPrivateInvariantsPolicy[consumer][methodSigs[i]] = policies[i];
             emit InvariantPolicySet(consumer, methodSigs[i], policies[i]);
@@ -476,7 +476,7 @@ contract Firewall is IFirewall, Ownable2Step {
     function _addPolicy(address consumer, bytes4 methodSig, address policy) internal {
         require(approvedPolicies[policy], "Firewall: policy not approved");
         address[] memory policies = subscribedPolicies[consumer][methodSig];
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             require(policy != policies[i], "Firewall: policy already exists");
         }
         subscribedPolicies[consumer][methodSig].push(policy);
@@ -491,7 +491,7 @@ contract Firewall is IFirewall, Ownable2Step {
      */
     function _removePolicy(address consumer, bytes4 methodSig, address policy) internal {
         address[] storage policies = subscribedPolicies[consumer][methodSig];
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             if (policy == policies[i]) {
                 policies[i] = policies[policies.length - 1];
                 policies.pop();
@@ -509,7 +509,7 @@ contract Firewall is IFirewall, Ownable2Step {
     function _addGlobalPolicy(address consumer, address policy) internal {
         require(approvedPolicies[policy], "Firewall: policy not approved");
         address[] memory policies = subscribedGlobalPolicies[consumer];
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             require(policy != policies[i], "Firewall: policy already exists");
         }
         subscribedGlobalPolicies[consumer].push(policy);
@@ -523,7 +523,7 @@ contract Firewall is IFirewall, Ownable2Step {
      */
     function _removeGlobalPolicy(address consumer, address policy) internal {
         address[] storage globalPolicies = subscribedGlobalPolicies[consumer];
-        for (uint i = 0; i < globalPolicies.length; i++) {
+        for (uint256 i = 0; i < globalPolicies.length; i++) {
             if (policy == globalPolicies[i]) {
                 globalPolicies[i] = globalPolicies[globalPolicies.length - 1];
                 globalPolicies.pop();
