+++ contracts/proxy/FirewallProxyIntercept.sol
@@ -100,7 +100,7 @@ contract FirewallProxyIntercept is TransparentUpgradeableProxy {
             return;
         }
 
-        uint value;
+        uint256 value;
         // We do this because msg.value can only be accessed in payable functions.
         assembly {
             value := callvalue()
